requirements.txt
Django>=5.2.5
djangorestframework>=3.14.0
django-filter>=23.0
django-cors-headers>=4.0.0
drf-nested-routers>=0.93.4 